refactor: Move matrix logic to matrix.ts
